refactor : mypage -> profle
